refactor (LottoTicketGenerator.java) : generate()의 매개변수가 자동과 수동이 서로 달라 오버라이딩 부가능해서 삭제
