--- src/main/java/lotto/domain/strategy/LottoTicketAutoGenerator.java
@@ -1,24 +1,0 @@
-package lotto.domain.strategy;
-
-import lotto.domain.item.LottoNumbers;
-import lotto.domain.item.LottoTicket;
-import lotto.domain.item.LottoTickets;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class LottoTicketAutoGenerator implements LottoTicketGenerator {
-
-    public LottoTicketAutoGenerator() {
-    }
-
-    @Override
-    public LottoTickets generate(int count) {
-        List<LottoTicket> result = new ArrayList<>();
-        for (int i = 0; i < count; i++) {
-            LottoNumbers numbers = LottoNumbers.createLottoNumbersUseInteger(LottoNumberAutoGenerator.generateLottoNumber());
-            result.add(new LottoTicket(numbers));
-        }
-        return new LottoTickets(result);
-    }
-}
