--- src/main/java/lotto/domain/strategy/LottoTicketPassivityGenerator.java
@@ -1,4 +1,0 @@
-package lotto.domain.strategy;
-
-public class LottoTicketPassivityGenerator {
-}
